Added Windows support for Docker executor.
This commit will introduce the changes to the Mesos Docker interfaces
required to use the Docker executor to run Windows Containers in the
Mesos agent. In particular, since Windows Containers use named pipes to
connect do the Docker host, rather than a socket, we introduce the
plumbing to default to a named pipe connection when invoking the
`docker` command.

This work constitutes the beginning of the end of the work that will
eventually result in Mesos support of Windows Containers.

This review is partially based on r/52364, which was the work of Daniel
Pravat.

Lastly, this review has a planned regression in MESOS-6816. In other
parts of the codebase, we copy all environment variables from the system
before launching a process, overwriting user specified environment
variables. This is not correct behavior, but a half-measure that is
necessary, because Windows does not inherit environment variables by
default. In this commit, we make this behavior uniform across all places
where we are creating a process, because it is better for behavior to be
consistent before we get around to fixing it. We do have concrete plans
to come back and resolve this issue.

Review: https://reviews.apache.org/r/56505