Windows: Enabled 495 passing tests on Agent builds.
This commit will introduce a sizeable chunk of the trivially-passing
Agent tests to the Windows build. Unlike #54618, which added test files
to the build only if they needed no modifications to compile and build
on Windows, here we add test files to the build only if they contain one
or more tests that either don't compile or don't pass. Hence, this
commit will partially address MESOS-6698, MESOS-6699, MESOS-6703,
MESOS-6705, MESOS-6706, MESOS-6707, MESOS-6709, and MESOS-6715.

In many of the files we added to the build, some of the tests either
don't compile, or don't pass. For tests that trivially compiled but
didn't pass, we use the `TEST_X_TEMP_DISABLED_ON_WINDOWS` macros. For
the tests that didn't compile, we simply `#ifdef` them out for now.

Review: https://reviews.apache.org/r/54677