Fixed Agent tests for the case when the credential is missing.
In Agent tests where `Clock::pause` is called and no credential is
passed into the Agent, the agent will currently fail to register, and
the test will hang and die.

The root cause is that, when no credential is provided to the Agent
(which is currently every case where `HAS_AUTHENTICATION` is not defined
by autotools or CMake), `Slave::detected` will use `delay` to schedule
Agent registration at a random time in the future. But, because the
clock is paused in these tests, it will never get executed. This issue
only manifests for the codepath when no credential is provided because,
when it is, the authentication codepath will attempt to register
immediately after authentication.

This issue currently blocks progress on MESOS-6698 and MESOS-6714.